TASK: Split utilities to a separate package
With this change all Utility classes that can be used
independently of Flow are moved to a separate package.
