+++ Neos.Flow.Utilities/Readme.rst
@@ -1,0 +1,16 @@
+--------------
+Flow Framework
+--------------
+
+.. note:: This repository is a **read-only subsplit** of a package that is part of the
+          Flow framework (learn more on `flow.typo3.org <http://flow.typo3.org/>`_).
+
+If you want to use the Flow framework, please have a look at the `Flow documentation
+<http://flowframework.readthedocs.org/en/stable/>`_
+
+Contribute
+----------
+
+If you want to contribute to the Flow framework, please have a look at
+https://github.com/neos/flow-development-collection - it is the repository
+used for development and all pull requests should go into it.
