+++ Neos.Flow.Utilities/Tests/Unit/PositionalArraySorterTest.php
@@ -2,7 +2,7 @@
 namespace TYPO3\Flow\Tests\Unit\Utility;
 
 /*
- * This file is part of the TYPO3.Flow package.
+ * This file is part of the Neos.Flow.Utilities package.
  *
  * (c) Contributors of the Neos Project - www.neos.io
  *
