+++ TYPO3.Flow/composer.json
@@ -28,7 +28,8 @@
         "symfony/dom-crawler": "~2.5.0",
         "symfony/console": "~2.0",
 
-        "neos/composer-plugin": "^1.0.1"
+        "neos/composer-plugin": "^1.0.1",
+        "neos/flow-utilities": "*"
     },
     "suggest": {
         "ext-curl": "To use the \\TYPO3\\Flow\\Http\\Client\\CurlEngine",
