+++ composer.json
@@ -7,8 +7,8 @@
     "type": "neos-package-collection",
     "require": {
         "php": ">=5.5.0",
-        "ext-zlib": "*",
         "ext-SPL": "*",
+        "ext-zlib": "*",
         "ext-json": ">=1.2.0",
         "ext-mbstring": "*",
         "ext-reflection": "*",
@@ -23,6 +23,7 @@
         "neos/composer-plugin": "^1.0.1"
     },
     "replace": {
+        "neos/flow-utilities": "self.version",
         "typo3/eel": "self.version",
         "typo3/flow": "self.version",
         "typo3/fluid": "self.version",
@@ -36,6 +37,7 @@
     },
     "autoload": {
         "psr-0": {
+            "TYPO3\\Flow\\Utility": "Neos.Flow.Utilities/Classes",
             "TYPO3\\Eel": "TYPO3.Eel/Classes",
             "TYPO3\\Flow": "TYPO3.Flow/Classes",
             "TYPO3\\Fluid": "TYPO3.Fluid/Classes",
